Returning HillClimbing due to bug with Greedy algorithm
Added isGeneral to time slot model to allow general time slots

Allowed for general time slots to be scheduled without sending them to
the algorithm functions. The current Bug with this is that the name for
that assignment on each group is NULL as no station is assigned.

Modified available time slots in station view to disallow opening stations
during general time slots

Added general time slots to dev initializer